consumerLens/requirements.txt
fastapi
uvicorn
pandas
openai
python-dotenv
pydantic